상속을 통한 확장
