+++ src/main/java/tobyspring/hellospring/payment/Client.java
@@ -1,0 +1,17 @@
+package tobyspring.hellospring.payment;
+
+import tobyspring.hellospring.payment.service.SimpleExRatePaymentService;
+import tobyspring.hellospring.payment.service.WebApiExRatePaymentService;
+import tobyspring.hellospring.payment.dao.Payment;
+import tobyspring.hellospring.payment.service.PaymentService;
+
+import java.io.IOException;
+import java.math.BigDecimal;
+
+public class Client {
+    public static void main(String[] args) throws IOException {
+        PaymentService paymentService = new SimpleExRatePaymentService();
+        Payment payment = paymentService.prepare(1L, "USD", BigDecimal.valueOf(1304));
+        System.out.println("payment = " + payment);
+    }
+}
