+++ src/main/java/tobyspring/hellospring/payment/service/PaymentService.java
@@ -1,44 +1,19 @@
 package tobyspring.hellospring.payment.service;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import tobyspring.hellospring.payment.dao.ExRateDate;
 import tobyspring.hellospring.payment.dao.Payment;
 
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.math.BigDecimal;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.time.LocalDateTime;
-import java.util.stream.Collectors;
 
-public class PaymentService {
+abstract public class PaymentService {
     public Payment prepare(Long orderId, String currency, BigDecimal foreignCurrencyAmount) throws IOException {
         BigDecimal exRate = getExRate(currency);
-
         BigDecimal convertedAmount = foreignCurrencyAmount.multiply(exRate);
-
         LocalDateTime validUntil = LocalDateTime.now().plusMinutes(30);
 
         return new Payment(orderId, currency, foreignCurrencyAmount, exRate, convertedAmount, validUntil);
     }
 
-    private BigDecimal getExRate(String currency) throws IOException {
-        URL url = new URL("https://open.er-api.com/v6/latest/" + currency);
-        HttpURLConnection urlConnection = (HttpURLConnection) url.openConnection();
-        BufferedReader br = new BufferedReader(new InputStreamReader(urlConnection.getInputStream()));
-        String response = br.lines().collect(Collectors.joining());
-        br.close();
-
-        ObjectMapper mapper = new ObjectMapper();
-        ExRateDate data = mapper.readValue(response, ExRateDate.class);
-        return data.rates().get("KRW");
-    }
-
-    public static void main(String[] args) throws IOException {
-        PaymentService paymentService = new PaymentService();
-        Payment payment = paymentService.prepare(1L, "USD", BigDecimal.valueOf(1304));
-        System.out.println("payment = " + payment);
-    }
+    protected abstract BigDecimal getExRate(String currency) throws IOException;
 }
