+++ src/main/java/tobyspring/hellospring/payment/service/WebApiExRatePaymentService.java
@@ -1,0 +1,28 @@
+package tobyspring.hellospring.payment.service;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import tobyspring.hellospring.payment.dao.ExRateDate;
+import tobyspring.hellospring.payment.service.PaymentService;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.math.BigDecimal;
+import java.net.HttpURLConnection;
+import java.net.URL;
+import java.util.stream.Collectors;
+
+public class WebApiExRatePaymentService extends PaymentService {
+    @Override
+    protected BigDecimal getExRate(String currency) throws IOException {
+        URL url = new URL("https://open.er-api.com/v6/latest/" + currency);
+        HttpURLConnection urlConnection = (HttpURLConnection) url.openConnection();
+        BufferedReader br = new BufferedReader(new InputStreamReader(urlConnection.getInputStream()));
+        String response = br.lines().collect(Collectors.joining());
+        br.close();
+
+        ObjectMapper mapper = new ObjectMapper();
+        ExRateDate data = mapper.readValue(response, ExRateDate.class);
+        return data.rates().get("KRW");
+    }
+}
